tableWithOdp: separated Table and GenerateValues components
