+++ webapp/app/assessment/fra/components/tableWithOdp/components/table/tableWithOdp.less
