+++ webapp/app/assessment/fra/components/tableWithOdp/index.js
@@ -1,0 +1,3 @@
+export { default } from './components/table/tableWithOdp'
+
+export { default as GenerateFraValuesControl } from './components/generateValues/generateValues'
