+++ webapp/app/assessment/fra/sections/extentOfForest/extentOfForestView.js
@@ -5,7 +5,7 @@ import * as R from 'ramda'
 import Icon from '@webapp/components/icon'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ChartWrapper from './chart/chartWrapper'
-import { GenerateFraValuesControl, TableWithOdp } from '../../components/tableWithOdp/tableWithOdp'
+import TableWithOdp, { GenerateFraValuesControl } from '@webapp/app/assessment/fra/components/tableWithOdp'
 import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
 import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
 import GeneralComments from '@webapp/components/description/generalComments'
